feat: refactor parse xml
