+++ src/utils/PLC/base-xml.ts
@@ -1,12 +1,9 @@
-import { ProjectState, RungState } from '@root/renderer/store/slices'
+import { ProjectState } from '@root/renderer/store/slices'
 import { BaseXml } from '@root/types/PLC/xml-data/base-diagram'
 import { create } from 'xmlbuilder2'
 
 import formatDate from '../formatDate'
-import { ilToXML } from './il-xml'
-import { ladderToXml } from './ladder-xml'
-import { parseInterface } from './pou-interface-xml'
-import { stToXML } from './st-xml'
+import { parsePousToXML } from './pou-xml'
 
 export const baseXmlStructure: BaseXml = {
   project: {
@@ -76,61 +73,8 @@ export const parseProjectToXML = (project: ProjectState) => {
    * Parse POUs
    */
   const pous = project.data.pous
-  pous.forEach((pou) => {
+  parsePousToXML(xmlResult, pous)
 
-    const interfaceResult = parseInterface(pou)
-
-    switch (pou.data.body.language) {
-      case 'il': {
-        const result = ilToXML(pou.data.body.value)
-        xmlResult.project.types.pous.pou.push({
-          '@name': pou.data.name,
-          '@pouType': pou.type === 'function-block' ? 'functionBlock' : pou.type,
-          interface: interfaceResult,
-          body: result.body,
-          documentation: {
-            'xhtml:p': {
-              $: pou.data.documentation,
-            },
-          },
-        })
-        return
-      }
-      case 'st': {
-        const result = stToXML(pou.data.body.value)
-        xmlResult.project.types.pous.pou.push({
-          '@name': pou.data.name,
-          '@pouType': pou.type === 'function-block' ? 'functionBlock' : pou.type,
-          interface: interfaceResult,
-          body: result.body,
-          documentation: {
-            'xhtml:p': {
-              $: pou.data.documentation,
-            },
-          },
-        })
-        return
-      }
-      case 'ld': {
-        const rungs = pou.data.body.value.rungs
-        const result = ladderToXml(rungs as RungState[])
-        xmlResult.project.types.pous.pou.push({
-          '@name': pou.data.name,
-          '@pouType': pou.type === 'function-block' ? 'functionBlock' : pou.type,
-          interface: interfaceResult,
-          body: result.body,
-          documentation: {
-            'xhtml:p': {
-              $: pou.data.documentation,
-            },
-          },
-        })
-        return
-      }
-      default:
-        return
-    }
-  })
 
   const doc = create(xmlResult)
   const xml = doc.end({ prettyPrint: true })
